refactor(core): move mapWithConcurrency to @appstrate/core
The bounded worker pool lived in apps/api/src/lib and was about to be
duplicated by the CLI. It has no API-specific dependency, so it moves to
core as the `@appstrate/core/map-with-concurrency` subpath, with its test.
The three API call sites import it from core; behaviour is unchanged.
